[vcpkg] Extract binary caching functionality into separate interface
